requirements.txt
# Core Flask Framework
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Limiter==4.0.0
alembic==1.16.5

# Configuration Management
python-decouple==3.8

# HTTP Client
requests==2.32.3
urllib3==2.2.1

# PDF Generation
PyPDF2==3.0.1
reportlab==4.2.5
xhtml2pdf==0.2.16

# Task Queue & Scheduling
celery==5.3.6
redis==5.0.3
APScheduler==3.11.0

# Security
cryptography==42.0.5
python-dotenv==1.0.1

# Production WSGI Server
gunicorn==21.2.0
gevent==24.2.1

# Monitoring & Logging
python-json-logger==2.0.7
psutil==5.9.8

# Testing (Development only)
pytest==8.1.1
pytest-cov==5.0.0

# Barcode Generation
python-barcode>=0.15.1
Pillow>=10.0.0

# Database
psycopg2-binary==2.9.9

# ===================================================================
# AI ASSISTANT DEPENDENCIES
# ===================================================================
# Choose ONE of the following LLM providers (or install multiple for flexibility):

# Option 1: Google Gemini 1.5 Flash (Recommended - CHEAPEST!)
# Cost: ~$0.075 per 1M tokens ($0.11/month for 50 queries/day)
google-generativeai>=0.3.0

# Option 2: OpenAI GPT-4o-mini (Fast & Affordable)
# Cost: ~$0.15 per 1M tokens ($0.23/month for 50 queries/day)
# openai>=1.12.0

# Option 3: Anthropic Claude 3.5 Haiku
# Cost: ~$0.25 per 1M tokens ($0.38/month for 50 queries/day)
# anthropic>=0.18.0

# Option 4: Ollama (Local LLM - FREE!)
# Requires Ollama installed locally: https://ollama.com
# Models: ministral-3:3b, deepseek-r1:8b, llama3.2:3b, etc.
ollama>=0.4.0